retrofited the page to allow any kind of downloads
