+++ app.go
@@ -213,9 +213,11 @@ func (a *App) GetUserHomeDirectory() string {
 
 func (a *App) ReadExportStatus() {
 	for {
+		//fmt.Println(" ")
 		exportStatus := a.exporter.GetExportStatus()
 
 		for _, item := range exportStatus.Feeds {
+			//fmt.Printf("> %s\n - %s - %d", "update-"+item.FeedName, item.FeedName, item.Counter)
 			runtime.EventsEmit(a.ctx, "update-"+item.FeedName, item)
 		}
